==> neuralcvd/preprocessing/0_decode_ukbb.ipynb <==
library(ukbtools)

my_ukb_data <- ukb_df("decoded", path = "/data/project/uk_bb/cvd/data/ukb_downloads/updated_showcase_43098/ukb_data")
df_field <- ukb_df_field("decoded", path = "/data/project/uk_bb/cvd/data/ukb_downloads/updated_showcase_43098/ukb_data")

arrow::write_feather(my_ukb_data, "/data/analysis/ag-reils/ag-reils-shared/cardioRS/data/1_decoded/ukb_data.feather")
arrow::write_feather(df_field, "/data/analysis/ag-reils/ag-reils-shared/cardioRS/data/1_decoded/ukb_data_field.feather")
